fix(images): read originals via presigned URL in preprocessing
Root cause of the 100%-failed backfill: the processor fetched originals with a
plain public fetch(image.url), but originals commonly live in a private bucket,
so every request returned HTTP 403 Forbidden (stage: fetch).

Fix: read the original through a credentialed presigned GET — the same
mechanism the "download original" route already uses for private storage. This
covers both private buckets and referer/hotlink protection (a presigned URL is
authenticated regardless).

- Extract download.ts's buildPresignedUrlForStorage into a shared
  server/lib/original-presign.ts (verbatim: key derivation from image.url,
  storage-folder handling, generatePresignedUrl). download.ts now imports it,
  so the download route and the pipeline share one proven implementation
  (zero drift).
- The processor resolves the key against the configured variant_storage
  backend (s3/r2). In a single-backend deployment — the common case — that is
  also where the originals live (download.ts + BE-4 read the same `bucket` /
  `r2_bucket` config key), so no per-image backend inference is needed.

Known limitation (not silent — surfaces as a fetch failure via the CLI's issue
output): if originals live in a *different* backend than variant_storage, or on
Open List, the presign targets the wrong bucket and that image fails at fetch.
Multi-backend support would need host-based backend inference (follow-up).
metadata-refresh.ts has the same latent private-original issue and could adopt
this helper later (follow-up).

Presign TTL is the generatePresignedUrl default (3600s), comfortably above the
30s fetch timeout.
